+++ src/components/atoms/InputLabel/InputLabel.tsx
@@ -17,14 +17,14 @@ const StyledInputLabel = styled.label<InputLabelProps & { theme: AppTheme }>`
   letter-spacing: 0;
   color: ${colors.greyDarkest};
   font-family: ${({ theme }) => theme.typography.primary};
-  font-weight: ${typography.weights.semiBold};
+  font-weight: ${({ theme }) => theme.input.labelFontWeight};
   font-size: ${typography.sizes.text.body};
+  line-height: ${({ theme }) => theme.input?.labelLineHeight};
 `;
 
 // TODO: Styleguidist to be able to locate styled components. See #147.
 export const InputLabel = (props: InputLabelProps) => {
   const theme = useThemeContext();
-
   return <StyledInputLabel {...props} theme={theme} />;
 };